move LensKit test utilities to lenskit.testing
