+++ conftest.py
@@ -16,8 +16,10 @@ from hypothesis import settings
 from pytest import fixture, skip
 
 from lenskit.parallel import ensure_parallel_init
+
+# bring common fixtures into scope
+from lenskit.testing import ml_100k, ml_ds, ml_ratings  # noqa: F401
 from lenskit.util.random import set_global_rng
-from lenskit.util.test import ml_100k, ml_ds, ml_ratings  # noqa: F401
 
 logging.getLogger("numba").setLevel(logging.INFO)
