+++ lenskit/tests/basic/test_popular.py
@@ -8,10 +8,8 @@ import pickle
 
 import pandas as pd
 
-import lenskit.util.test as lktu
 from lenskit.basic import PopScorer
 from lenskit.data.items import ItemList
-from lenskit.util.test import ml_ds, ml_ratings  # noqa: F401
 
 simple_df = pd.DataFrame(
     {"item": [1, 1, 2, 3], "user": [10, 12, 10, 13], "rating": [4.0, 3.0, 5.0, 2.0]}
