+++ lenskit/tests/basic/test_random.py
@@ -18,7 +18,7 @@ from hypothesis import HealthCheck, assume, given, settings
 
 from lenskit.basic.random import RandomSelector
 from lenskit.data.items import ItemList
-from lenskit.util.test import scored_lists
+from lenskit.testing import scored_lists
 
 _log = logging.getLogger(__name__)
