+++ lenskit/tests/basic/test_topn.py
@@ -15,11 +15,9 @@ import hypothesis.extra.numpy as nph
 import hypothesis.strategies as st
 from hypothesis import given, settings
 
-import lenskit.util.test as lktu
 from lenskit.basic import PopScorer
 from lenskit.basic.topn import TopNRanker
 from lenskit.data.items import ItemList
-from lenskit.util.test import ml_ds, ml_ratings  # noqa: F401
 
 _log = logging.getLogger(__name__)
