+++ lenskit/tests/data/test_dataset_stats.py
@@ -15,7 +15,7 @@ from pytest import approx
 
 from lenskit.data import Dataset
 from lenskit.data.tables import NumpyUserItemTable, TorchUserItemTable
-from lenskit.util.test import ml_ds, ml_ratings  # noqa: F401
+from lenskit.testing import ml_ds, ml_ratings  # noqa: F401
 
 
 def test_item_stats(ml_ratings: pd.DataFrame, ml_ds: Dataset):
