+++ lenskit/tests/eval/test_rank_mrr.py
@@ -11,7 +11,7 @@ from pytest import approx, mark
 
 from lenskit.data import ItemList
 from lenskit.metrics.ranking import RecipRank
-from lenskit.util.test import demo_recs  # noqa: F401
+from lenskit.testing import demo_recs  # noqa: F401
 
 
 def _test_rr(items, rel, **kw):
