+++ lenskit/tests/eval/test_rank_precision.py
@@ -11,7 +11,7 @@ from pytest import approx
 
 from lenskit.data import ItemList
 from lenskit.metrics.ranking import Precision
-from lenskit.util.test import demo_recs  # noqa: F401
+from lenskit.testing import demo_recs  # noqa: F401
 
 
 def _test_prec(items, rel, **k):
