+++ lenskit/tests/eval/test_rank_rbp.py
@@ -16,7 +16,7 @@ from pytest import approx, mark
 from lenskit.data import ItemList
 from lenskit.metrics import call_metric
 from lenskit.metrics.ranking import RBP
-from lenskit.util.test import demo_recs  # noqa: F401
+from lenskit.testing import demo_recs  # noqa: F401
 
 _log = logging.getLogger(__name__)
