+++ lenskit/tests/utils/test_matrix_rows.py
@@ -17,7 +17,7 @@ from hypothesis import HealthCheck, given, settings
 from pytest import approx
 
 from lenskit.math.sparse import normalize_sparse_rows
-from lenskit.util.test import sparse_tensors
+from lenskit.testing import sparse_tensors
 
 _log = logging.getLogger(__name__)
